# reports/verify-61c0f29/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Microservice Code Studio",
    "description": "Microservice Code Studio - Autonomous Java 21 / Spring Boot 3.x Generator",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tests/synthesize": {
      "post": {
        "tags": [
          "Tests & Self-Repair"
        ],
        "summary": "Synthesize hybrid unit and integration test suites",
        "description": "Synthesizes Mockito unit tests, @WebMvcTest controller tests, and @SpringBootTest context tests.",
        "operationId": "synthesize_test_suites_api_v1_tests_synthesize_post",
        "parameters": [
          {
            "name": "X-LLM-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Llm-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestSynthesisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestSynthesisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/analyze": {
      "post": {
        "tags": [
          "Tests & Self-Repair"
        ],
        "summary": "Analyze build logs, test results, and static code compliance",
        "description": "Parses compiler logs and test failure stack traces into structured FailureDiagnostics,\nand runs static constitutional compliance checks on the source files.",
        "operationId": "analyze_code_and_failures_api_v1_tests_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/repair": {
      "post": {
        "tags": [
          "Tests & Self-Repair"
        ],
        "summary": "Plan and apply an autonomous surgical self-repair patch",
        "description": "Plans and applies a surgical method/block patch bounded strictly by 3 iterations (Constitution Principle V).",
        "operationId": "execute_repair_iteration_api_v1_tests_repair_post",
        "parameters": [
          {
            "name": "X-LLM-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Llm-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepairExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairIterationRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/repairs": {
      "get": {
        "tags": [
          "Tests & Self-Repair"
        ],
        "summary": "Retrieve complete repair history and diffs for a session",
        "description": "Returns the history of repair attempts, applied patches, and diff summaries.",
        "operationId": "get_session_repairs_api_v1_sessions__sessionId__repairs_get",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepairHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionId}/manual-repair": {
      "post": {
        "tags": [
          "Tests & Self-Repair"
        ],
        "summary": "Submit developer manual code edit or AI prompt hint to resolve a BLOCKED session",
        "description": "Allows the developer to provide a manual code fix or natural language hint to unblock a session.",
        "operationId": "submit_manual_repair_api_v1_sessions__sessionId__manual_repair_post",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sessionid"
            }
          },
          {
            "name": "X-LLM-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Llm-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualRepairRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualRepairResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodeAnalysisRequest": {
        "properties": {
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "rawBuildLogs": {
            "type": "string",
            "title": "Rawbuildlogs",
            "default": ""
          },
          "sourceFiles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sourcefiles"
          }
        },
        "type": "object",
        "title": "CodeAnalysisRequest"
      },
      "CodeAnalysisResponse": {
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "title": "Haserrors"
          },
          "constitutionalCompliant": {
            "type": "boolean",
            "title": "Constitutionalcompliant"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/FailureDiagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "violations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Violations"
          }
        },
        "type": "object",
        "required": [
          "hasErrors",
          "constitutionalCompliant",
          "diagnostics"
        ],
        "title": "CodeAnalysisResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualRepairRequest": {
        "properties": {
          "filePath": {
            "type": "string",
            "title": "Filepath",
            "description": "File to modify"
          },
          "modifiedCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modifiedcode",
            "description": "Full replacement source or patched snippet"
          },
          "guidanceHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidancehint",
            "description": "Natural language direction for the repair agent"
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "required": [
          "filePath"
        ],
        "title": "ManualRepairRequest"
      },
      "ManualRepairResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "diagnosticsResolved": {
            "type": "boolean",
            "title": "Diagnosticsresolved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "status",
          "message"
        ],
        "title": "ManualRepairResponse"
      },
...
      "RepairExecutionRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "iterationNumber": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Iterationnumber"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/FailureDiagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "sourceFiles": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sourcefiles"
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "iterationNumber",
          "diagnostics"
        ],
        "title": "RepairExecutionRequest"
      },
      "RepairHistoryResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "totalIterations": {
            "type": "integer",
            "title": "Totaliterations"
          },
          "finalState": {
            "type": "string",
            "title": "Finalstate"
          },
          "iterations": {
            "items": {
              "$ref": "#/components/schemas/RepairIterationRecord"
            },
            "type": "array",
            "title": "Iterations"
          },
          "canRetryManually": {
            "type": "boolean",
            "title": "Canretrymanually",
            "default": false
          },
          "currentBlockedDiagnostic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureDiagnostic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "totalIterations",
          "finalState",
          "iterations"
        ],
        "title": "RepairHistoryResponse"
      },
      "RepairIterationRecord": {
        "properties": {
          "iterationNumber": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Iterationnumber"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/FailureDiagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "patchesApplied": {
            "items": {
              "$ref": "#/components/schemas/CodeRepairPatch"
            },
            "type": "array",
            "title": "Patchesapplied"
          },
          "passedTestsBefore": {
            "type": "integer",
            "title": "Passedtestsbefore",
            "default": 0
          },
          "failedTestsBefore": {
            "type": "integer",
            "title": "Failedtestsbefore",
            "default": 0
          },
          "passedTestsAfter": {
            "type": "integer",
            "title": "Passedtestsafter",
            "default": 0
          },
          "failedTestsAfter": {
            "type": "integer",
            "title": "Failedtestsafter",
            "default": 0
          },
          "diffSummary": {
            "type": "string",
            "title": "Diffsummary",
            "description": "Unified diff of all modifications in this iteration",
            "default": ""
          },
          "durationSeconds": {
            "type": "number",
            "title": "Durationseconds",
            "default": 0.0
          },
          "outcome": {
            "$ref": "#/components/schemas/RepairOutcome"
          }
        },
        "type": "object",
        "required": [
          "iterationNumber",
          "diagnostics",
          "patchesApplied",
          "outcome"
        ],
        "title": "RepairIterationRecord"
      },
      "RepairOutcome": {
        "type": "string",
        "enum": [
          "SUCCESS",
          "FAILED_CONTINUE",
          "FAILED_BLOCKED"
        ],
        "title": "RepairOutcome"
      },
...
      "TestSynthesisRequest": {
        "properties": {
          "specId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specid"
          },
          "blueprint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blueprint"
          },
          "testTypes": {
            "items": {
              "$ref": "#/components/schemas/TestType"
            },
            "type": "array",
            "title": "Testtypes"
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "title": "TestSynthesisRequest"
      },
      "TestSynthesisResponse": {
        "properties": {
          "specId": {
            "type": "string",
            "title": "Specid"
          },
          "serviceName": {
            "type": "string",
            "title": "Servicename"
          },
          "packageName": {
            "type": "string",
            "title": "Packagename"
          },
          "suites": {
            "items": {
              "$ref": "#/components/schemas/TestSuiteDefinition"
            },
            "type": "array",
            "title": "Suites"
          },
          "totalTestCases": {
            "type": "integer",
            "title": "Totaltestcases"
          }
        },
        "type": "object",
        "required": [
          "specId",
          "serviceName",
          "packageName",
          "suites",
          "totalTestCases"
        ],
        "title": "TestSynthesisResponse"
      },
      "TestType": {
        "type": "string",
        "enum": [
          "UNIT",
          "INTEGRATION_WEB",
          "INTEGRATION_DB"
        ],
        "title": "TestType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}